+++ custom_components/axeos_ha_integration/button.py
@@ -2,7 +2,7 @@
 
 import logging
 
-from homeassistant.components.button import ButtonEntity
+from homeassistant.components.button import ButtonEntity, ButtonEntityDescription
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
@@ -13,6 +13,12 @@ from .const import DOMAIN
 from .models import AxeOSConfigEntry
 
 _LOGGER = logging.getLogger(__name__)
+
+RESTART_BUTTON_DESCRIPTION = ButtonEntityDescription(
+    key="restart",
+    translation_key="restart",
+    icon="mdi:restart",
+)
 
 
 async def async_setup_entry(
@@ -56,8 +62,7 @@ class AxeOSRestartButton(CoordinatorEntity, ButtonEntity):
         self.host_id = host_id
         self.api = api
 
-        self._attr_name = "Restart"
-        self._attr_icon = "mdi:restart"
+        self.entity_description = RESTART_BUTTON_DESCRIPTION
         self._attr_unique_id = f"{host_id}_restart_button"
 
     async def async_press(self) -> None: